refactor(compiler-shared): consolidate shared types
