+++ packages/compiler-shared/src/types.ts
@@ -1,18 +1,10 @@
 /**
- * Data-shape contract shared by the Panda compiler bindings — the serialized
- * surface `@pandacss/compiler` (native) and `@pandacss/compiler-wasm` (browser)
- * both speak. Pure types, no runtime, no internal imports (leaf module).
+ * Data-shape contract shared by the Panda compiler bindings.
+ * Pure types only; keep runtime helpers in sibling modules.
  */
 
 export * from './types/diagnostics'
 export * from './types/extraction'
-export * from './types/tokens'
-export * from './types/recipes'
-export * from './types/css'
-export * from './types/spec'
-export * from './types/codegen'
-export * from './types/parsing'
+export * from './types/output'
 export * from './types/config'
-export * from './types/callbacks'
-export * from './types/build-info'
 export * from './types/compiler'
