--- packages/compiler-shared/src/types/callbacks.ts
@@ -1,38 +1,0 @@
-import type { ColorMixResult, RawToken } from './tokens'
-
-export interface TransformArgs {
-  token: ((path: string) => string | undefined) & { raw: (path: string) => RawToken | undefined }
-  raw: unknown
-  utils: {
-    colorMix(value: string): ColorMixResult
-  }
-}
-
-export interface PatternHelpers {
-  map(value: unknown, fn: (value: any) => any): unknown
-  isCssUnit(value: unknown): boolean
-  isCssVar(value: unknown): boolean
-  isCssFunction(value: unknown): boolean
-}
-
-export type UtilityValuesTheme = (category: string) => Record<string, string> | undefined
-export type UtilityValuesCallback = (theme: UtilityValuesTheme) => Record<string, string> | string[] | undefined
-export type UtilityTransformCallback = (value: string, args: TransformArgs) => unknown
-export type PatternTransformCallback = (props: Record<string, any>, helpers: PatternHelpers) => unknown
-export type PatternDefaultValuesCallback = (props: Record<string, any>) => unknown
-export interface SourceTransformArgs {
-  filePath: string
-  content: string
-  original?: string
-}
-export type SourceTransformCallback = (args: SourceTransformArgs) => string | void
-
-export interface ProjectCallbackMap {
-  'utility.transform': Record<string, UtilityTransformCallback>
-  'utility.values': Record<string, UtilityValuesCallback>
-  'pattern.transform': Record<string, PatternTransformCallback>
-  'pattern.defaultValues': Record<string, PatternDefaultValuesCallback>
-  'parser:before': Record<string, SourceTransformCallback>
-}
-
-export type ProjectCallbacks = Partial<ProjectCallbackMap>
